requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
pillow
streamlit
tensorflow-cpu
scikit-learn
